requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.0
psycopg2-binary==2.9.9
python-dotenv==1.1.1
bcrypt==4.1.2